User cereberus's normalized method
This normalizes the data against the configschema.yaml schema. For
example, to embed default values if a field isn't explicitly set.

This is demonstrated by the "optional" field in dialog fields.

A by-product is that placeholder is now defaulting to an empty string,
so the code needs to work around that.