Add escape_yaml_doublequoted Jinja filter
This filter is useful for escaping content that is intended to be
inserted as a double-quoted string in YAML. Double-quoted YAML strings
are useful because they can potentially hold any conceivable string.
However, double quotes and backslashes within the content need to be
escaped, of course. This Jinja filter makes that possible for
templatekit users.

Includes tests.