Separate out the periodic CI run
When applying the same pattern to Safir, I realized that a weekly
periodic run might refresh the tox environment cache frequently
enough that it would always be considered current and the tox
environments would never be recreated, thus defeating the point of
testing against the latest version.

Follow the Safir setup and instead separate the periodic run into
a different workflow that doesn't use the tox environment cache.
Share a pip cache across all the workflows to still minimize
downloads.